docs(ledger)+fix(harness): DISC-006 RESOLVED — 96000/96000 sample-for-sample under canonical gate
The comparator's i16 quantization now frounds the product to match
sample_to_i16's f32 multiply (the JS f64 product straddled where the
CLI's f32 product tied: -734.49998 vs -734.5 -> -734 vs -735). With
matched binaries at --packet-frames 1: 96,000/96,000 samples
identical, all tapI/tapM/tapC hashes bit-equal. Default-route
divergence remains documented; canonical mode is the opt-in parity
mechanism (frankentts-rc-disc006-wasm-divergence-4oqv closed).

+++ site/harness/browser.mjs
@@ -317,7 +317,12 @@ try {
         // +infinity, which disagrees with Rust's f32::round on negative halves
         // and manufactured phantom 1-LSB diffs at exact .5 boundaries.
         const ours = pcm.slice(0, n).map((v) => {
-          const scaled = v * 32767;
+          // Math.fround replicates the CLI's f32 arithmetic: sample_to_i16
+          // computes v * 32767.0 in f32 (the product rounds to f32 BEFORE the
+          // integer rounding), so the f64 product here must be f32-rounded
+          // first or exact-half products disagree by 1 LSB
+          // (ftts_core::audio::sample_to_i16; frankentts-uuac).
+          const scaled = Math.fround(v * 32767);
           const rounded = Math.sign(scaled) * Math.round(Math.abs(scaled));
           return Math.max(-32768, Math.min(32767, rounded));
         });
